Remove Formatter namespace
