+++ .rubocop.yml
@@ -107,8 +107,8 @@ RSpec/MultipleDescribes:
   Exclude:
     - 'spec/reek/ast/sexp_extensions_spec.rb'
     - 'spec/reek/code_comment_spec.rb'
-    - 'spec/reek/report/formatter/location_formatter_spec.rb'
-    - 'spec/reek/report/formatter/progress_formatter_spec.rb'
+    - 'spec/reek/report/location_formatter_spec.rb'
+    - 'spec/reek/report/progress_formatter_spec.rb'
 
 # FIXME: Update specs to avoid offenses
 RSpec/MultipleExpectations:
