+++ lib/reek/spec/should_reek.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../examiner'
-require_relative '../report/formatter/simple_warning_formatter'
+require_relative '../report/simple_warning_formatter'
 
 module Reek
   module Spec
@@ -23,7 +23,7 @@ module Reek
       end
 
       def failure_message_when_negated
-        rpt = Report::Formatter::SimpleWarningFormatter.new.format_list(examiner.smells)
+        rpt = Report::SimpleWarningFormatter.new.format_list(examiner.smells)
         "Expected no smells, but got:\n#{rpt}"
       end
