+++ spec/reek/report/text_report_spec.rb
@@ -1,15 +1,15 @@
 require_relative '../../spec_helper'
 require_lib 'reek/examiner'
 require_lib 'reek/report/text_report'
-require_lib 'reek/report/formatter/heading_formatter'
-require_lib 'reek/report/formatter/simple_warning_formatter'
+require_lib 'reek/report/heading_formatter'
+require_lib 'reek/report/simple_warning_formatter'
 require 'rainbow'
 
 RSpec.describe Reek::Report::TextReport do
   let(:report_options) do
     {
-      warning_formatter: Reek::Report::Formatter::SimpleWarningFormatter.new,
-      heading_formatter: Reek::Report::Formatter::QuietHeadingFormatter
+      warning_formatter: Reek::Report::SimpleWarningFormatter.new,
+      heading_formatter: Reek::Report::QuietHeadingFormatter
     }
   end
   let(:instance) { described_class.new report_options }
