+++ spec/reek/report_spec.rb
@@ -10,19 +10,19 @@ RSpec.describe Reek::Report do
 
   describe '.location_formatter' do
     it 'returns the correct class' do
-      expect(described_class.location_formatter(:plain)).to eq Reek::Report::Formatter::BlankLocationFormatter
+      expect(described_class.location_formatter(:plain)).to eq Reek::Report::BlankLocationFormatter
     end
   end
 
   describe '.heading_formatter' do
     it 'returns the correct class' do
-      expect(described_class.heading_formatter(:quiet)).to eq Reek::Report::Formatter::QuietHeadingFormatter
+      expect(described_class.heading_formatter(:quiet)).to eq Reek::Report::QuietHeadingFormatter
     end
   end
 
   describe '.warning_formatter_class' do
     it 'returns the correct class' do
-      expect(described_class.warning_formatter_class(:simple)).to eq Reek::Report::Formatter::SimpleWarningFormatter
+      expect(described_class.warning_formatter_class(:simple)).to eq Reek::Report::SimpleWarningFormatter
     end
   end
 end
